spell check, more comments, and fix close { func
spell check shortcut: alt-F5 to enable, alt-F6 to disable (english only,
so far).
when closing an open {, now we check in current line for any *unmatched*
open bracket, even if after it there one or more pairs of closed
brackets (a check that was missing, i.e. if there was a pair of {} in
the same line, the next bracket would go to a new line, even if before
that there was an { that had no corresponding }). This was annoying in
Perl because of access to hashtables within hashtables...