.pandoc is now .cmk;texvim scope var change;spell
The .pandoc extension is now .cmk, after commonmark. The file type must
still be pandoc, for vim-pandoc-syntax to function, but it makes much
more sense for the relevant extension to be .cmk, given that pandoc is
just the parser for the commonmark markdown flavour.

In ftplugin/tex.vim, script variables are now either buffer or local
variables. This means that, for example, if I disable compile on save,
and open a new buffer on the same windows, compile on save remains
disabled on the original buffer.

vimrc now has spell checking normal maps:

- <leader>sse: switch on spell checking in english
- <leader>ssp: switch on spell checking in portuguese
- <leader>ssn: switch off spell checking